fix(docs): fix 404 flash — swizzle @theme/NotFound so SPA and static match
src/pages/404.tsx only controls the static 404.html served by GitHub
Pages. When the Docusaurus SPA boots on that page it renders
@theme/NotFound (the default "Page Not Found" component), overwriting
the custom content with the generic Docusaurus 404.

Fix: create src/theme/NotFound.tsx which overrides @theme/NotFound for
the SPA catch-all route. src/pages/404.tsx now re-exports the same
component so both the static file and the SPA route are identical.
